+++ test/commands/post_mortem_test.rb
@@ -56,7 +56,7 @@ module Byebug
 
     ['restart', 'frame', 'quit', 'edit', 'info', 'irb', 'source', 'help',
      'var class', 'list', 'method', 'kill', 'eval', 'set', 'save', 'show',
-     'thread list'].each do |cmd|
+     'thread list', 'up', 'where', 'down'].each do |cmd|
       define_method "test_#{cmd}_is_permitted_in_post_mortem_mode" do
         enter 'set post_mortem', "#{cmd}", 'set no_postmortem'
         class_name = cmd.gsub(/(^| )\w/) { |b| b[-1, 1].upcase } + 'Command'